+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to insta and cargo-insta are documented here.
   moved since the time of the snapshot creation. (#220)
 - `cargo insta test` now returns non zero status code when snapshots
   are left for review. (#222)
+- Assertion failures now mention `cargo insta test`. (#223)
 
 ## 1.13.0